Overwrite the .sdirs file with cat instead of mv
I am using dropbox to sync my .sdirs file amongst my workstations, and I
symlink the ~/.sdirs to the actual sdirs file, which is stored
elsewhere. Using mv (or mv -f, as was also suggested in another pull
request) blows away the symlink every time I add or remove a bookmark.

This solution allows for the .sdirs file to be overwritten, but will
keep the file's location intact.